test rm multi_devices_graph_pass
