+++ paddle/fluid/framework/CMakeLists.txt
@@ -889,8 +889,7 @@ cc_library(
        eager_deletion_pass
        buffer_shared_inplace_op_pass
        buffer_shared_cross_op_memory_reuse_pass
-       inplace_addto_op_pass
-       set_reader_device_info_utils)
+       inplace_addto_op_pass)
 
 cc_library(
   executor_cache
