+++ paddle/fluid/framework/details/CMakeLists.txt
@@ -20,8 +20,6 @@ set(op_handle_srcs
     var_handle.cc
     op_handle_base.cc
     scale_loss_grad_op_handle.cc
-    fetch_op_handle.cc
-    fetch_async_op_handle.cc
     share_tensor_buffer_functor.cc
     computation_op_handle.cc
     share_tensor_buffer_op_handle.cc
